refactoring TokenType
