+++ contracts/liquidity_book/lb_pair/Cargo.toml
@@ -19,7 +19,7 @@ backtraces = ["cosmwasm-std/backtraces"]
 
 [dependencies]
 shade-protocol = { version = "0.1.0", path = "../../../packages/shade_protocol", features = [
-  "liquidity_book","storage_plus","lb-libraries",
+  "liquidity_book","storage_plus","lb-libraries","swap",
 ] }
 cosmwasm-std = { package = "secret-cosmwasm-std", version = "=1.1.11", features = ["iterator"] }
 cosmwasm-storage = { package = "secret-cosmwasm-storage", version = "=1.1.11", features = ["iterator"] }
