+++ contracts/liquidity_book/rust-toolchain
@@ -1,0 +1,1 @@
+1.69.0
