+++ packages/multi_test/Cargo.toml
@@ -65,7 +65,7 @@ admin = { version = "0.2.0", path = "../../contracts/admin", optional = true }
 mock_stkd = { version = "0.1.0", package = "mock_stkd_derivative", path = "../../contracts/mock/mock_stkd_derivative", optional = true }
 mock_sienna = { version = "0.1.0", package = "mock_sienna_pair", path = "../../contracts/mock/mock_sienna_pair", optional = true }
 snip20_migration = { version = "0.1.0", path = "../../contracts/snip20_migration", optional = true }
-shade-protocol = { path = "../shade_protocol", features = ["multi-test"] }
+shade-protocol = { path = "../shade_protocol", features = ["multi-test","swap"] }
 # shadeswap-shared = {  version = "0.1.0", path = "../../packages/shadeswap_shared", features = []  }
 anyhow = "1"
