+++ packages/shade_protocol/src/contract_interfaces/swap/core/display.rs
@@ -1,5 +1,14 @@
-use super::TokenPair;
+use super::{TokenPair, TokenType};
 use std::fmt::{Display, Formatter, Result};
+
+impl Display for TokenType {
+    fn fmt(&self, f: &mut Formatter) -> Result {
+        match self {
+            TokenType::NativeToken { denom, .. } => write!(f, "{}", denom),
+            TokenType::CustomToken { contract_addr, .. } => write!(f, "{}", contract_addr),
+        }
+    }
+}
 
 impl Display for TokenPair {
     fn fmt(&self, f: &mut Formatter) -> Result {
