+++ packages/shade_protocol/src/contract_interfaces/swap/mod.rs
@@ -2,7 +2,6 @@ use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
 pub mod core;
-
 pub mod amm_pair;
 pub mod factory;
 pub mod router;
