+++ packages/shade_protocol/src/utils/liquidity_book/types.rs
@@ -10,7 +10,7 @@ use ethnum::U256;
 pub use crate::utils::liquidity_book::math::{
     liquidity_configurations::LiquidityConfigurations, tree_math::TreeUint24,
 };
-pub use crate::utils::liquidity_book::tokens::TokenType;
+use crate::swap::core::TokenType;
 
 pub type Bytes32 = [u8; 32];
